Move fake and fixture objects out of production code
