+++ tests/Unit/Storage/Reaction/ReportingStorageReactionTest.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace Haspadar\Piqule\Tests\Unit\Storage\Reaction;
 
 use Haspadar\Piqule\Storage\Reaction\ReportingStorageReaction;
-use Haspadar\Piqule\Tests\Unit\Fake\Output\FakeOutput;
+use Haspadar\Piqule\Tests\Fake\Output\FakeOutput;
 use PHPUnit\Framework\Attributes\Test;
 use PHPUnit\Framework\TestCase;
